Add support for configuring the ignored files via the cli
