+++ flow-typed/minimatch.js
@@ -1,0 +1,3 @@
+declare module 'minimatch' {
+  declare export default function minimatch(filePath: string, glob: string): boolean;
+}
