+++ flow-typed/yargs.js
@@ -1,0 +1,5 @@
+declare module 'yargs' {
+  declare export default {
+    argv: Object,
+  };
+}
